Bumped package

+++ src/main/java/edu/iga/adi/sm/SolverConfiguration.java
@@ -98,7 +98,7 @@ public class SolverConfiguration {
 
     @Parameter(names = {"--terrain-scale"})
     @Builder.Default
-    private int scale = 1; // 100
+    private double scale = 1; // 100
 
     @Parameter(names = {"--terrain-x-offset"})
     @Builder.Default
